span.hh: improve old MSVC support
The main thing here is that we keep a check for
    !std::is_convertible<End, size_t>
in the is_citer_pair trait, which allows the iter+size constructor to be
selected over the iter+end constructor.

This reverts the following commits, as they were only necessary because
iter+end ctor was being erroneously selected:
73f1e19aba2a043ebe352043e4aa5d398dc4cb02
c584e76132e7fb65f566e57dbbf2be7ee3678a64
f810bf3a1607ae58ebfd13ee87efd00ba753d76e

This also ports a different expression-SFINAE method from the following
commit. This should work all the way back to MSVC 19.0 (vs2015).
DankRank/vhvc@6167204ff0fe97ffab6eca94dcaa34029834bbaa

+++ src/libromdata/Other/EXE_NE.cpp
@@ -579,7 +579,7 @@ int EXEPrivate::addFields_NE_Entry(void)
 			return -ENOENT;
 
 		while (*p) {
-			size_t len = *p++;
+			uint8_t len = *p++;
 			if (p + len + 2 >= end) // next length byte >= end
 				return -ENOENT;
 			vhvc::span<const char> name(reinterpret_cast<const char *>(p), len);
